Parking Spot and Lot

+++ src/ticketing/Ticket.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <cstddef>
 #include <optional>
 #include <chrono>
 #include <string>
@@ -7,14 +8,14 @@ typedef std::chrono::system_clock::time_point c_time;
 
 class Ticket {
 private:
-    int id;
+    size_t id;
     std::string vehicleTag;
     std::chrono::system_clock::time_point entryTime;
     std::optional<std::chrono::system_clock::time_point> exitTime;
     bool paid;
 public:
     Ticket(
-        int id,
+        size_t id,
         std::string vehicleTag,
         c_time entryTime,
         std::optional<std::chrono::system_clock::time_point> exitTime) :
